refactor(history): split SQLite snapshot diagnostics
