+++ crates/ctx-history-capture/src/provider_sources/sqlite_source/snapshot.rs
@@ -27,8 +27,12 @@ struct LogicalSnapshotHooks<
 }
 
 fn ignore_snapshot_path(_: &Path) {}
+mod acquisition;
+mod backup_handle;
+mod certification;
 mod copy_progress;
 mod scratch;
+mod source_copy;
 #[cfg(test)]
 mod test_api;
 #[cfg(test)]
@@ -45,7 +49,9 @@ pub(super) use test_api::{
     open_root_handle_sqlite_source_snapshot_for_test, run_online_backup_with_deadline_for_test,
 };
 
+pub(super) use acquisition::{close_private_snapshot_directory, close_private_sqlite_connection};
 use copy_progress::{copy_sqlite_member_with_progress, report_source_family_copy_progress};
+use {acquisition::*, backup_handle::*, certification::*, source_copy::*};
 
 /// Retains an approved parent-directory handle together with the pathname that
 /// stock SQLite is allowed to open beneath it.
@@ -71,7 +77,6 @@ pub(crate) fn open_root_handle_sqlite_source_snapshot(
         || {},
     )
 }
-
 #[cfg(test)]
 pub(super) fn open_root_handle_sqlite_source_snapshot_with_policy(
     authority: &SqliteSourceDirectoryAuthority,
@@ -648,377 +653,6 @@ where
     })
 }
 
-struct OnlineBackupSource {
-    connection: Connection,
-    copied_source_directory: Option<TempDir>,
-    copied_source_bytes: u64,
-    artifact: SqliteArtifactKind,
-}
-
-fn close_online_backup_source(
-    source: OnlineBackupSource,
-    pinned: bool,
-    copied_pages: u64,
-    copied_bytes: u64,
-) -> SqliteSourceAccessResult<SqliteCleanupStatus> {
-    let OnlineBackupSource {
-        connection,
-        copied_source_directory,
-        artifact,
-        ..
-    } = source;
-    let end_snapshot = if pinned {
-        end_pinned_read_snapshot(&connection).map_err(|source| {
-            SqliteSourceAccessError::CleanupUnavailable {
-                operation: "ending the private SQLite source-copy transaction",
-                source: Box::new(source),
-            }
-            .with_diagnostic(
-                SqliteFailurePhase::Cleanup,
-                artifact,
-                copied_pages,
-                copied_bytes,
-                SqliteCleanupStatus::Failed,
-            )
-        })
-    } else {
-        Ok(())
-    };
-    let close_connection = close_private_sqlite_connection(
-        connection,
-        "closing the private SQLite source-copy connection",
-        artifact,
-        copied_pages,
-        copied_bytes,
-    );
-    let cleanup_status = if copied_source_directory.is_some() {
-        SqliteCleanupStatus::Succeeded
-    } else {
-        SqliteCleanupStatus::NotRequired
-    };
-    let close_directory = copied_source_directory.map_or(Ok(()), |directory| {
-        close_private_snapshot_directory(directory, artifact, copied_pages, copied_bytes)
-    });
-    match (end_snapshot, close_connection, close_directory) {
-        (_, _, Err(error)) | (_, Err(error), Ok(())) | (Err(error), Ok(()), Ok(())) => Err(error),
-        (Ok(()), Ok(()), Ok(())) => Ok(cleanup_status),
-    }
-}
-
-fn acquire_online_backup_source<E>(
-    data_root: &Path,
-    snapshot_context: &Arc<SqliteSourceSnapshotContext>,
-    family: &SqliteSourceFamily,
-    evidence: &SqliteFamilyEvidence,
-    copy_hooks: (impl FnOnce(), impl FnOnce(&Path)),
-    scratch_limit: u64,
-    report_progress: &mut impl FnMut(SourceBackedCurrentSourceProgress) -> Result<(), E>,
-) -> Result<OnlineBackupSource, SqliteSourceProgressError<E>> {
-    let (after_database_copy, after_private_source_copy) = copy_hooks;
-    let committed_wal = evidence.wal.as_ref().is_some_and(|state| state.length != 0);
-    if !committed_wal {
-        #[cfg(target_os = "linux")]
-        if immutable_procfd_available(family.database.file()) {
-            return Ok(OnlineBackupSource {
-                connection: open_immutable_main(&family.database)?,
-                copied_source_directory: None,
-                copied_source_bytes: 0,
-                artifact: SqliteArtifactKind::ProviderDatabase,
-            });
-        }
-
-        let copied_bytes =
-            enforce_snapshot_copy_bounds_with_limit(family, evidence, scratch_limit)?;
-        let (snapshot_directory, snapshot_path, integrity) =
-            copy_sqlite_family_to_ctx_with_progress(
-                data_root,
-                family,
-                evidence,
-                after_database_copy,
-                report_progress,
-            )?;
-        if let Err(error) = snapshot_context.record_source_bytes_copied(copied_bytes) {
-            return match close_private_snapshot_directory(
-                snapshot_directory,
-                SqliteArtifactKind::PrivateSourceCopy,
-                0,
-                copied_bytes,
-            ) {
-                Ok(()) => Err(error.into()),
-                Err(cleanup) => Err(cleanup.into()),
-            };
-        }
-        after_private_source_copy(&snapshot_path);
-        if let Err(error) = certify_private_source_copy(&snapshot_path, &integrity, copied_bytes) {
-            return match close_private_snapshot_directory(
-                snapshot_directory,
-                SqliteArtifactKind::PrivateSourceCopy,
-                0,
-                copied_bytes,
-            ) {
-                Ok(()) => Err(error.into()),
-                Err(cleanup) => Err(cleanup.into()),
-            };
-        }
-        return finish_copied_online_backup_source(
-            snapshot_directory,
-            snapshot_path,
-            copied_bytes,
-            "opening the exact copied logical-backup source",
-        )
-        .map_err(Into::into);
-    }
-
-    let copied_bytes = enforce_snapshot_copy_bounds_with_limit(family, evidence, scratch_limit)?;
-    let (snapshot_directory, snapshot_path, integrity) = copy_sqlite_family_to_ctx_with_progress(
-        data_root,
-        family,
-        evidence,
-        after_database_copy,
-        report_progress,
-    )?;
-    if let Err(error) = snapshot_context.record_source_bytes_copied(copied_bytes) {
-        return match close_private_snapshot_directory(
-            snapshot_directory,
-            SqliteArtifactKind::PrivateSourceCopy,
-            0,
-            copied_bytes,
-        ) {
-            Ok(()) => Err(error.into()),
-            Err(cleanup) => Err(cleanup.into()),
-        };
-    }
-    after_private_source_copy(&snapshot_path);
-    if let Err(error) = certify_private_source_copy(&snapshot_path, &integrity, copied_bytes) {
-        return match close_private_snapshot_directory(
-            snapshot_directory,
-            SqliteArtifactKind::PrivateSourceCopy,
-            0,
-            copied_bytes,
-        ) {
-            Ok(()) => Err(error.into()),
-            Err(cleanup) => Err(cleanup.into()),
-        };
-    }
-    finish_copied_online_backup_source(
-        snapshot_directory,
-        snapshot_path,
-        copied_bytes,
-        "opening the ctx-owned online-backup source",
-    )
-    .map_err(Into::into)
-}
-
-fn finish_copied_online_backup_source(
-    snapshot_directory: TempDir,
-    snapshot_path: PathBuf,
-    copied_bytes: u64,
-    operation: &'static str,
-) -> SqliteSourceAccessResult<OnlineBackupSource> {
-    let result = Connection::open_with_flags(
-        &snapshot_path,
-        OpenFlags::SQLITE_OPEN_READ_ONLY
-            | OpenFlags::SQLITE_OPEN_NO_MUTEX
-            | OpenFlags::SQLITE_OPEN_PRIVATE_CACHE
-            | OpenFlags::SQLITE_OPEN_NOFOLLOW,
-    )
-    .map_err(|source| {
-        sqlite_error(operation, source).with_diagnostic(
-            SqliteFailurePhase::SourceValidation,
-            SqliteArtifactKind::PrivateSourceCopy,
-            0,
-            copied_bytes,
-            SqliteCleanupStatus::NotRequired,
-        )
-    });
-    match result {
-        Ok(connection) => Ok(OnlineBackupSource {
-            connection,
-            copied_source_directory: Some(snapshot_directory),
-            copied_source_bytes: copied_bytes,
-            artifact: SqliteArtifactKind::ProviderDatabase,
-        }),
-        Err(error) => match close_private_snapshot_directory(
-            snapshot_directory,
-            SqliteArtifactKind::PrivateSourceCopy,
-            0,
-            copied_bytes,
-        ) {
-            Ok(()) => Err(error),
-            Err(cleanup) => Err(cleanup),
-        },
-    }
-}
-
-fn certify_private_source_copy(
-    snapshot_path: &Path,
-    integrity: &CopiedFamilyIntegrity,
-    copied_bytes: u64,
-) -> SqliteSourceAccessResult<()> {
-    let database_digest = private_snapshot_file_digest(snapshot_path, copied_bytes)?;
-    let wal_path = snapshot_path.with_file_name("source.sqlite-wal");
-    let wal_digest = if integrity.wal_digest.is_some() {
-        Some(private_snapshot_file_digest(&wal_path, copied_bytes)?)
-    } else {
-        None
-    };
-    if database_digest == integrity.database_digest && wal_digest == integrity.wal_digest {
-        Ok(())
-    } else {
-        Err(SqliteSourceAccessError::SqliteControl {
-            operation: "certifying the exact private SQLite source-family copy",
-            code: ffi::SQLITE_CORRUPT,
-        }
-        .with_diagnostic(
-            SqliteFailurePhase::SourceValidation,
-            SqliteArtifactKind::PrivateSourceCopy,
-            0,
-            copied_bytes,
-            SqliteCleanupStatus::NotRequired,
-        ))
-    }
-}
-
-fn private_snapshot_file_digest(
-    path: &Path,
-    copied_bytes: u64,
-) -> SqliteSourceAccessResult<[u8; 32]> {
-    let mut file = File::open(path).map_err(|source| {
-        SqliteSourceAccessError::ScratchIoUnavailable {
-            operation: "opening a private SQLite source component for certification",
-            path: path.to_path_buf(),
-            source,
-        }
-        .with_diagnostic(
-            SqliteFailurePhase::SourceValidation,
-            SqliteArtifactKind::PrivateSourceCopy,
-            0,
-            copied_bytes,
-            SqliteCleanupStatus::NotRequired,
-        )
-    })?;
-    let mut digest = Sha256::new();
-    let mut buffer = [0_u8; SQLITE_COPY_BUFFER_BYTES];
-    loop {
-        let read = file.read(&mut buffer).map_err(|source| {
-            SqliteSourceAccessError::ScratchIoUnavailable {
-                operation: "reading a private SQLite source component for certification",
-                path: path.to_path_buf(),
-                source,
-            }
-            .with_diagnostic(
-                SqliteFailurePhase::SourceValidation,
-                SqliteArtifactKind::PrivateSourceCopy,
-                0,
-                copied_bytes,
-                SqliteCleanupStatus::NotRequired,
-            )
-        })?;
-        if read == 0 {
-            break;
-        }
-        digest.update(&buffer[..read]);
-    }
-    Ok(digest.finalize().into())
-}
-
-#[derive(Clone, Copy)]
-struct OnlineBackupBounds {
-    page_count: u64,
-    page_size: u64,
-    bytes: u64,
-}
-
-fn enforce_online_backup_bounds(
-    connection: &Connection,
-    path: &Path,
-    scratch_limit: u64,
-) -> SqliteSourceAccessResult<OnlineBackupBounds> {
-    let page_count: i64 = connection
-        .pragma_query_value(None, "page_count", |row| row.get(0))
-        .map_err(|source| sqlite_error("reading online-backup page count", source))?;
-    let page_size: i64 = connection
-        .pragma_query_value(None, "page_size", |row| row.get(0))
-        .map_err(|source| sqlite_error("reading online-backup page size", source))?;
-    let page_count =
-        u64::try_from(page_count).map_err(|_| SqliteSourceAccessError::SnapshotUnavailable {
-            reason: "the provider SQLite page count is negative".to_owned(),
-        })?;
-    let page_size =
-        u64::try_from(page_size).map_err(|_| SqliteSourceAccessError::SnapshotUnavailable {
-            reason: "the provider SQLite page size is negative".to_owned(),
-        })?;
-    let bytes = page_count.checked_mul(page_size).ok_or_else(|| {
-        SqliteSourceAccessError::SnapshotTooLarge {
-            path: path.to_path_buf(),
-            length: u64::MAX,
-            maximum: scratch_limit,
-        }
-    })?;
-    if bytes > scratch_limit {
-        return Err(SqliteSourceAccessError::SnapshotTooLarge {
-            path: path.to_path_buf(),
-            length: bytes,
-            maximum: scratch_limit,
-        });
-    }
-    Ok(OnlineBackupBounds {
-        page_count,
-        page_size,
-        bytes,
-    })
-}
-
-fn certify_sqlite_snapshot(
-    connection: &Connection,
-    bounds: OnlineBackupBounds,
-    phase: SqliteFailurePhase,
-    artifact: SqliteArtifactKind,
-    copied_pages: u64,
-    copied_bytes: u64,
-) -> SqliteSourceAccessResult<SqliteValidationMeasurement> {
-    let started = Instant::now();
-    let deadline = started
-        .checked_add(SQLITE_CERTIFICATION_DEADLINE)
-        .ok_or_else(|| SqliteSourceAccessError::SnapshotUnavailable {
-            reason: "the SQLite certification deadline overflowed".to_owned(),
-        })?;
-    connection.progress_handler(
-        SQLITE_CERTIFICATION_PROGRESS_OPS,
-        Some(move || Instant::now() >= deadline),
-    );
-    let result = connection.query_row("PRAGMA quick_check(1)", [], |row| row.get::<_, String>(0));
-    connection.progress_handler(0, None::<fn() -> bool>);
-    let result = result.map_err(|source| {
-        sqlite_error("certifying the pinned SQLite snapshot", source).with_diagnostic(
-            phase,
-            artifact,
-            copied_pages,
-            copied_bytes,
-            SqliteCleanupStatus::NotRequired,
-        )
-    })?;
-    if result != "ok" {
-        return Err(SqliteSourceAccessError::SqliteControl {
-            operation: "certifying the pinned SQLite snapshot",
-            code: ffi::SQLITE_CORRUPT,
-        }
-        .with_diagnostic(
-            phase,
-            artifact,
-            copied_pages,
-            copied_bytes,
-            SqliteCleanupStatus::NotRequired,
-        ));
-    }
-    Ok(SqliteValidationMeasurement {
-        pages: bounds.page_count,
-        bytes: bounds.bytes,
-        #[cfg(test)]
-        elapsed_ms: u64::try_from(started.elapsed().as_millis()).unwrap_or(u64::MAX),
-    })
-}
-
 fn online_backup_to_ctx<E>(
     data_root: &Path,
     source: &Connection,
@@ -1325,464 +959,4 @@ fn online_backup_progress(
     progress.snapshot_bytes_completed = Some(completed_bytes);
     progress.snapshot_bytes_total = Some(bounds.bytes);
     Ok(progress)
-}
-
-#[cfg(test)]
-fn run_online_backup_until(
-    source: &Connection,
-    destination: &Connection,
-    deadline: Instant,
-) -> SqliteSourceAccessResult<()> {
-    let backup = unsafe {
-        ffi::sqlite3_backup_init(
-            destination.handle(),
-            c"main".as_ptr(),
-            source.handle(),
-            c"main".as_ptr(),
-        )
-    };
-    if backup.is_null() {
-        return Err(SqliteSourceAccessError::SqliteControl {
-            operation: "initializing the logical SQLite online backup",
-            code: unsafe { ffi::sqlite3_extended_errcode(destination.handle()) },
-        });
-    }
-    let mut backup = OnlineBackupHandle(Some(backup));
-    loop {
-        if Instant::now() >= deadline {
-            return Err(online_backup_deadline_error());
-        }
-        let code = unsafe {
-            ffi::sqlite3_backup_step(backup.pointer(), SQLITE_ONLINE_BACKUP_PAGES_PER_STEP)
-        };
-        if Instant::now() >= deadline {
-            return Err(online_backup_deadline_error());
-        }
-        match code {
-            ffi::SQLITE_DONE => break,
-            ffi::SQLITE_OK => continue,
-            ffi::SQLITE_BUSY | ffi::SQLITE_LOCKED if Instant::now() < deadline => {
-                thread::sleep(Duration::from_millis(10));
-            }
-            code => {
-                return Err(SqliteSourceAccessError::SqliteControl {
-                    operation: "copying the pinned logical SQLite snapshot",
-                    code,
-                });
-            }
-        }
-    }
-    let code = backup.finish();
-    if code == ffi::SQLITE_OK {
-        Ok(())
-    } else {
-        Err(SqliteSourceAccessError::SqliteControl {
-            operation: "finishing the logical SQLite online backup",
-            code,
-        })
-    }
-}
-
-fn online_backup_deadline_error() -> SqliteSourceAccessError {
-    SqliteSourceAccessError::SnapshotUnavailable {
-        reason: "the logical SQLite online backup exceeded its five-minute deadline".to_owned(),
-    }
-}
-
-struct OnlineBackupHandle(Option<*mut ffi::sqlite3_backup>);
-
-impl OnlineBackupHandle {
-    fn pointer(&self) -> *mut ffi::sqlite3_backup {
-        self.0.unwrap_or(ptr::null_mut())
-    }
-
-    fn finish(&mut self) -> i32 {
-        self.0.take().map_or(ffi::SQLITE_OK, |backup| unsafe {
-            ffi::sqlite3_backup_finish(backup)
-        })
-    }
-}
-
-impl Drop for OnlineBackupHandle {
-    fn drop(&mut self) {
-        let _ = self.finish();
-    }
-}
-
-fn end_pinned_read_snapshot(connection: &Connection) -> SqliteSourceAccessResult<()> {
-    clear_snapshot_authorizer(connection)?;
-    connection
-        .execute_batch("ROLLBACK")
-        .map_err(|source| sqlite_error("ending the provider online-backup snapshot", source))
-}
-
-struct AcquiredSqliteConnection {
-    connection: Connection,
-    #[cfg(test)]
-    strategy: SqliteSourceSnapshotStrategy,
-    #[cfg(test)]
-    copied_bytes: u64,
-    snapshot_directory: Option<TempDir>,
-    snapshot_activity: SqliteSourceSnapshotActivity,
-}
-
-struct CopiedFamilyIntegrity {
-    database_digest: [u8; 32],
-    wal_digest: Option<[u8; 32]>,
-}
-
-impl AcquiredSqliteConnection {
-    fn cleanup(self) -> SqliteSourceAccessResult<()> {
-        let artifact = if self.snapshot_directory.is_some() {
-            SqliteArtifactKind::PrivateSourceCopy
-        } else {
-            SqliteArtifactKind::PrivateScratch
-        };
-        let close = close_private_sqlite_connection(
-            self.connection,
-            "closing a rejected SQLite source snapshot",
-            artifact,
-            0,
-            0,
-        );
-        let cleanup = self.snapshot_directory.map_or(Ok(()), |directory| {
-            close_private_snapshot_directory(directory, artifact, 0, 0)
-        });
-        drop(self.snapshot_activity);
-        match (close, cleanup) {
-            (_, Err(error)) | (Err(error), Ok(())) => Err(error),
-            (Ok(()), Ok(())) => Ok(()),
-        }
-    }
-}
-
-fn acquire_sqlite_connection(
-    data_root: &Path,
-    snapshot_context: &Arc<SqliteSourceSnapshotContext>,
-    family: &SqliteSourceFamily,
-    evidence: &SqliteFamilyEvidence,
-    after_database_copy: impl FnOnce(),
-) -> SqliteSourceAccessResult<AcquiredSqliteConnection> {
-    if family.wal.is_none() && family.shared_memory.is_none() {
-        #[cfg(target_os = "linux")]
-        if immutable_procfd_available(family.database.file()) {
-            let connection = open_immutable_main(&family.database)?;
-            let snapshot_activity = match snapshot_context
-                .record_open(SqliteSourceSnapshotStrategy::ImmutableMain, 0)
-            {
-                Ok(activity) => activity,
-                Err(error) => {
-                    return match close_private_sqlite_connection(
-                        connection,
-                        "closing an untracked immutable SQLite snapshot",
-                        SqliteArtifactKind::PrivateScratch,
-                        0,
-                        0,
-                    ) {
-                        Ok(()) => Err(error),
-                        Err(cleanup) => Err(cleanup),
-                    };
-                }
-            };
-            return Ok(AcquiredSqliteConnection {
-                connection,
-                #[cfg(test)]
-                strategy: SqliteSourceSnapshotStrategy::ImmutableMain,
-                #[cfg(test)]
-                copied_bytes: 0,
-                snapshot_directory: None,
-                snapshot_activity,
-            });
-        }
-    }
-
-    let copied_bytes = enforce_snapshot_copy_bounds(family, evidence)?;
-    let (snapshot_directory, snapshot_path) =
-        copy_sqlite_family_to_ctx(data_root, family, evidence, after_database_copy)?;
-    let connection = Connection::open_with_flags(
-        &snapshot_path,
-        OpenFlags::SQLITE_OPEN_READ_ONLY
-            | OpenFlags::SQLITE_OPEN_NO_MUTEX
-            | OpenFlags::SQLITE_OPEN_PRIVATE_CACHE
-            | OpenFlags::SQLITE_OPEN_NOFOLLOW,
-    )
-    .map_err(|source| sqlite_error("opening the ctx-owned provider snapshot", source));
-    let connection = match connection {
-        Ok(connection) => connection,
-        Err(error) => {
-            return match close_private_snapshot_directory(
-                snapshot_directory,
-                SqliteArtifactKind::PrivateSourceCopy,
-                0,
-                copied_bytes,
-            ) {
-                Ok(()) => Err(error),
-                Err(cleanup) => Err(cleanup),
-            };
-        }
-    };
-    if let Err(error) = snapshot_context.record_source_bytes_copied(copied_bytes) {
-        let close = close_private_sqlite_connection(
-            connection,
-            "closing an unaccounted SQLite source snapshot",
-            SqliteArtifactKind::PrivateSourceCopy,
-            0,
-            copied_bytes,
-        );
-        let cleanup = close_private_snapshot_directory(
-            snapshot_directory,
-            SqliteArtifactKind::PrivateSourceCopy,
-            0,
-            copied_bytes,
-        );
-        return match (close, cleanup) {
-            (_, Err(cleanup)) | (Err(cleanup), Ok(())) => Err(cleanup),
-            (Ok(()), Ok(())) => Err(error),
-        };
-    }
-    let snapshot_activity = match snapshot_context
-        .record_open(SqliteSourceSnapshotStrategy::CopiedFamily, copied_bytes)
-    {
-        Ok(activity) => activity,
-        Err(error) => {
-            let close = close_private_sqlite_connection(
-                connection,
-                "closing an untracked SQLite source snapshot",
-                SqliteArtifactKind::PrivateSourceCopy,
-                0,
-                copied_bytes,
-            );
-            let cleanup = close_private_snapshot_directory(
-                snapshot_directory,
-                SqliteArtifactKind::PrivateSourceCopy,
-                0,
-                copied_bytes,
-            );
-            return match (close, cleanup) {
-                (_, Err(cleanup)) | (Err(cleanup), Ok(())) => Err(cleanup),
-                (Ok(()), Ok(())) => Err(error),
-            };
-        }
-    };
-    Ok(AcquiredSqliteConnection {
-        connection,
-        #[cfg(test)]
-        strategy: SqliteSourceSnapshotStrategy::CopiedFamily,
-        #[cfg(test)]
-        copied_bytes,
-        snapshot_directory: Some(snapshot_directory),
-        snapshot_activity,
-    })
-}
-
-#[cfg(target_os = "linux")]
-fn immutable_procfd_available(database: &File) -> bool {
-    PathBuf::from(format!("/proc/self/fd/{}", database.as_raw_fd())).exists()
-}
-
-#[cfg(target_os = "linux")]
-fn open_immutable_main(database: &SqliteFamilyMember) -> SqliteSourceAccessResult<Connection> {
-    let procfd_path = PathBuf::from(format!("/proc/self/fd/{}", database.file().as_raw_fd()));
-    let mut uri = Url::from_file_path(&procfd_path).map_err(|()| {
-        SqliteSourceAccessError::SnapshotUnavailable {
-            reason: "the retained SQLite main handle cannot be represented as a file URI"
-                .to_owned(),
-        }
-    })?;
-    uri.query_pairs_mut()
-        .append_pair("mode", "ro")
-        .append_pair("immutable", "1");
-    Connection::open_with_flags(
-        uri.as_str(),
-        OpenFlags::SQLITE_OPEN_READ_ONLY
-            | OpenFlags::SQLITE_OPEN_NO_MUTEX
-            | OpenFlags::SQLITE_OPEN_PRIVATE_CACHE
-            | OpenFlags::SQLITE_OPEN_URI,
-    )
-    .map_err(|source| sqlite_error("opening the retained immutable provider database", source))
-}
-
-fn enforce_snapshot_copy_bounds(
-    family: &SqliteSourceFamily,
-    evidence: &SqliteFamilyEvidence,
-) -> SqliteSourceAccessResult<u64> {
-    enforce_snapshot_copy_bounds_with_limit(family, evidence, SQLITE_SNAPSHOT_MAX_TOTAL_BYTES)
-}
-
-fn enforce_snapshot_copy_bounds_with_limit(
-    family: &SqliteSourceFamily,
-    evidence: &SqliteFamilyEvidence,
-    scratch_limit: u64,
-) -> SqliteSourceAccessResult<u64> {
-    // The family total is authoritative: main and WAL may consume any share,
-    // but every observed byte from both members must fit and be copied.
-    let mut total = evidence.database.length;
-    match (family.wal.as_ref(), evidence.wal.as_ref()) {
-        (Some(_wal), Some(state)) => {
-            total = total.checked_add(state.length).ok_or_else(|| {
-                SqliteSourceAccessError::SnapshotTooLarge {
-                    path: family.database.path.clone(),
-                    length: u64::MAX,
-                    maximum: scratch_limit,
-                }
-            })?;
-        }
-        (None, None) => {}
-        _ => return Err(SqliteSourceAccessError::SourceChanged),
-    }
-    if total > scratch_limit {
-        return Err(SqliteSourceAccessError::SnapshotTooLarge {
-            path: family.database.path.clone(),
-            length: total,
-            maximum: scratch_limit,
-        });
-    }
-    Ok(total)
-}
-
-fn copy_sqlite_family_to_ctx(
-    data_root: &Path,
-    family: &SqliteSourceFamily,
-    evidence: &SqliteFamilyEvidence,
-    after_database_copy: impl FnOnce(),
-) -> SqliteSourceAccessResult<(TempDir, PathBuf)> {
-    match copy_sqlite_family_to_ctx_with_progress(
-        data_root,
-        family,
-        evidence,
-        after_database_copy,
-        &mut |_| Ok::<(), std::convert::Infallible>(()),
-    ) {
-        Ok((directory, path, _)) => Ok((directory, path)),
-        Err(SqliteSourceProgressError::Source(error)) => Err(error),
-        Err(SqliteSourceProgressError::Progress(never)) => match never {},
-    }
-}
-
-fn copy_sqlite_family_to_ctx_with_progress<E>(
-    data_root: &Path,
-    family: &SqliteSourceFamily,
-    evidence: &SqliteFamilyEvidence,
-    after_database_copy: impl FnOnce(),
-    report_progress: &mut impl FnMut(SourceBackedCurrentSourceProgress) -> Result<(), E>,
-) -> Result<(TempDir, PathBuf, CopiedFamilyIntegrity), SqliteSourceProgressError<E>> {
-    let total_bytes = enforce_snapshot_copy_bounds(family, evidence)?;
-    let mut completed_bytes = 0;
-    let mut last_reported_bytes = 0;
-    report_source_family_copy_progress(report_progress, completed_bytes, total_bytes)?;
-    let directory = create_snapshot_directory(data_root, "provider-sqlite-snapshot-")?;
-    let snapshot_path = directory.path().join("source.sqlite");
-    let operation = (|| {
-        let database_digest = copy_sqlite_member_with_progress(
-            &family.database,
-            &snapshot_path,
-            evidence.database.length,
-            &mut completed_bytes,
-            &mut last_reported_bytes,
-            total_bytes,
-            report_progress,
-        )?;
-        after_database_copy();
-        family.revalidate(evidence)?;
-        let wal_digest = match (family.wal.as_ref(), evidence.wal.as_ref()) {
-            (Some(wal), Some(state)) => Some(copy_sqlite_member_with_progress(
-                wal,
-                &directory.path().join("source.sqlite-wal"),
-                state.length,
-                &mut completed_bytes,
-                &mut last_reported_bytes,
-                total_bytes,
-                report_progress,
-            )?),
-            (None, None) => None,
-            _ => return Err(SqliteSourceAccessError::SourceChanged.into()),
-        };
-        if completed_bytes != total_bytes {
-            return Err(SqliteSourceAccessError::SourceChanged.into());
-        }
-        family.revalidate(evidence)?;
-        Ok(CopiedFamilyIntegrity {
-            database_digest,
-            wal_digest,
-        })
-    })();
-    let integrity = match operation {
-        Ok(integrity) => integrity,
-        Err(error) => {
-            return match close_private_snapshot_directory(
-                directory,
-                SqliteArtifactKind::PrivateSourceCopy,
-                0,
-                completed_bytes,
-            ) {
-                Ok(()) => Err(error),
-                Err(cleanup) => Err(cleanup.into()),
-            }
-        }
-    };
-    // SHM is lock coordination, not provider content. Copying it would retain
-    // volatile reader marks. Stock SQLite rebuilds it only in this ctx-owned
-    // directory from the certified DB/WAL pair.
-    Ok((directory, snapshot_path, integrity))
-}
-
-fn create_snapshot_directory(data_root: &Path, prefix: &str) -> SqliteSourceAccessResult<TempDir> {
-    let staging_root = data_root.join("tmp").join("provider-sqlite");
-    create_private_directory_all(&staging_root).map_err(|source| {
-        SqliteSourceAccessError::ScratchIoUnavailable {
-            operation: "creating the private provider SQLite staging root",
-            path: staging_root.clone(),
-            source,
-        }
-    })?;
-    let directory = tempfile::Builder::new()
-        .prefix(prefix)
-        .tempdir_in(&staging_root)
-        .map_err(|source| SqliteSourceAccessError::ScratchIoUnavailable {
-            operation: "creating a private provider SQLite snapshot",
-            path: staging_root,
-            source,
-        })?;
-    Ok(directory)
-}
-
-pub(super) fn close_private_snapshot_directory(
-    directory: TempDir,
-    artifact: SqliteArtifactKind,
-    copied_pages: u64,
-    copied_bytes: u64,
-) -> SqliteSourceAccessResult<()> {
-    let path = directory.path().to_path_buf();
-    directory.close().map_err(|source| {
-        SqliteSourceAccessError::ScratchIoUnavailable {
-            operation: "removing a ctx-owned SQLite snapshot directory",
-            path,
-            source,
-        }
-        .with_diagnostic(
-            SqliteFailurePhase::Cleanup,
-            artifact,
-            copied_pages,
-            copied_bytes,
-            SqliteCleanupStatus::Failed,
-        )
-    })
-}
-
-pub(super) fn close_private_sqlite_connection(
-    connection: Connection,
-    operation: &'static str,
-    artifact: SqliteArtifactKind,
-    copied_pages: u64,
-    copied_bytes: u64,
-) -> SqliteSourceAccessResult<()> {
-    connection.close().map_err(|(_, source)| {
-        SqliteSourceAccessError::ScratchSqliteUnavailable { operation, source }.with_diagnostic(
-            SqliteFailurePhase::Cleanup,
-            artifact,
-            copied_pages,
-            copied_bytes,
-            SqliteCleanupStatus::Failed,
-        )
-    })
 }
