+++ crates/ctx-history-capture/src/provider_sources/sqlite_source/tests.rs
@@ -35,6 +35,9 @@ use super::{
     SqliteSourceSnapshotStrategy, SQLITE_SHM_MAX_BYTES, SQLITE_SNAPSHOT_MAX_TOTAL_BYTES,
 };
 
+mod copy_races;
+mod corruption_cleanup;
+mod diagnostics;
 mod path_safety;
 mod scratch;
 
@@ -148,145 +151,6 @@ fn directory_names(path: &Path) -> BTreeSet<OsString> {
         .unwrap()
         .map(|entry| entry.unwrap().file_name())
         .collect()
-}
-
-fn assert_revalidation_resource_unavailable(error: SqliteSourceAccessError) {
-    assert!(error.is_systemic_resource_failure());
-    assert!(matches!(
-        error,
-        SqliteSourceAccessError::ResourceUnavailable { .. }
-    ));
-}
-
-#[cfg(unix)]
-#[test]
-fn sqlite_authority_revalidation_preserves_descriptor_exhaustion() {
-    let path = Path::new("/provider/provider.sqlite");
-    for code in [libc::EMFILE, libc::ENFILE] {
-        let error = map_revalidation_io_error(
-            io::Error::from_raw_os_error(code),
-            "retaining SQLite authority in the regression test",
-            path,
-        );
-        match &error {
-            SqliteSourceAccessError::ResourceUnavailable {
-                operation,
-                path: error_path,
-                source,
-            } => {
-                assert_eq!(
-                    *operation,
-                    "retaining SQLite authority in the regression test"
-                );
-                assert_eq!(error_path, path);
-                assert_eq!(source.raw_os_error(), Some(code));
-            }
-            other => panic!("unexpected revalidation error: {other:?}"),
-        }
-        assert_revalidation_resource_unavailable(error);
-    }
-}
-
-#[test]
-fn sqlite_authority_revalidation_preserves_portable_resource_exhaustion() {
-    let path = Path::new("provider.sqlite");
-    for kind in [
-        io::ErrorKind::OutOfMemory,
-        io::ErrorKind::StorageFull,
-        io::ErrorKind::QuotaExceeded,
-    ] {
-        let error = map_revalidation_io_error(
-            io::Error::from(kind),
-            "retaining SQLite authority in the portable regression test",
-            path,
-        );
-        assert!(matches!(
-            &error,
-            SqliteSourceAccessError::ResourceUnavailable { source, .. }
-                if source.kind() == kind
-        ));
-        assert_revalidation_resource_unavailable(error);
-    }
-}
-
-#[test]
-fn sqlite_primary_resource_codes_are_exhaustive_for_control_and_rusqlite_errors() {
-    for code in [
-        ffi::SQLITE_FULL,
-        ffi::SQLITE_NOMEM,
-        ffi::SQLITE_IOERR,
-        ffi::SQLITE_CANTOPEN,
-        ffi::SQLITE_PERM,
-        ffi::SQLITE_READONLY,
-    ] {
-        let control = SqliteSourceAccessError::SqliteControl {
-            operation: "classifying a SQLite control failure",
-            code,
-        };
-        assert!(
-            control.is_systemic_resource_failure(),
-            "control code {code}"
-        );
-        let rusqlite = SqliteSourceAccessError::Sqlite {
-            operation: "classifying a rusqlite failure",
-            source: rusqlite::Error::SqliteFailure(ffi::Error::new(code), None),
-        };
-        assert!(
-            rusqlite.is_systemic_resource_failure(),
-            "rusqlite code {code}"
-        );
-    }
-    for code in [
-        ffi::SQLITE_BUSY,
-        ffi::SQLITE_LOCKED,
-        ffi::SQLITE_CORRUPT,
-        ffi::SQLITE_NOTADB,
-    ] {
-        let error = SqliteSourceAccessError::Sqlite {
-            operation: "classifying a non-resource SQLite failure",
-            source: rusqlite::Error::SqliteFailure(ffi::Error::new(code), None),
-        };
-        assert!(!error.is_systemic_resource_failure(), "SQLite code {code}");
-    }
-}
-
-#[cfg(unix)]
-#[test]
-fn raw_io_resource_codes_are_systemic_without_revalidation_rewriting() {
-    for code in [
-        libc::ENOSPC,
-        libc::EMFILE,
-        libc::ENFILE,
-        libc::ENOMEM,
-        libc::EDQUOT,
-    ] {
-        let error = SqliteSourceAccessError::Io {
-            operation: "reading a SQLite resource fixture",
-            path: PathBuf::from("provider.sqlite"),
-            source: io::Error::from_raw_os_error(code),
-        };
-        assert!(error.is_systemic_resource_failure(), "raw OS code {code}");
-    }
-}
-
-#[test]
-fn sqlite_authority_revalidation_keeps_mutation_as_source_changed() {
-    assert!(matches!(
-        map_revalidation_error(SqliteSourceAccessError::ConnectionIdentityMismatch),
-        SqliteSourceAccessError::SourceChanged
-    ));
-    assert!(matches!(
-        map_revalidation_error(SqliteSourceAccessError::SourceChanged),
-        SqliteSourceAccessError::SourceChanged
-    ));
-    assert!(matches!(
-        map_revalidation_io_error(
-            io::Error::from(io::ErrorKind::NotFound),
-            "reopening mutated SQLite authority in the regression test",
-            Path::new("provider.sqlite"),
-        ),
-        SqliteSourceAccessError::SourceChanged
-    ));
 }
 
 #[test]
@@ -1169,185 +1033,6 @@ fn logical_online_backup_is_one_private_snapshot_without_provider_content_or_nam
 }
 
 #[test]
-fn malformed_provider_is_source_local_and_is_not_retried_as_a_full_copy() {
-    let temp = tempfile::tempdir().unwrap();
-    let database = temp.path().join("provider.sqlite");
-    let _writer = create_persistent_wal(&database);
-    let mut database_file = OpenOptions::new().write(true).open(&database).unwrap();
-    database_file.seek(SeekFrom::Start(0)).unwrap();
-    database_file.write_all(&[0_u8; 100]).unwrap();
-    database_file.sync_all().unwrap();
-    drop(database_file);
-    let wal = database.with_file_name("provider.sqlite-wal");
-    let mut wal_file = OpenOptions::new().write(true).open(&wal).unwrap();
-    wal_file.seek(SeekFrom::Start(0)).unwrap();
-    wal_file.write_all(&[0_u8; 32]).unwrap();
-    wal_file.sync_all().unwrap();
-    drop(wal_file);
-    let expected_copy_bytes =
-        fs::metadata(&database).unwrap().len() + fs::metadata(&wal).unwrap().len();
-    let before = directory_file_bytes(temp.path());
-    let parent = retain_parent(temp.path());
-
-    let error = parent
-        .open_logical_online_backup_snapshot(OsStr::new("provider.sqlite"))
-        .unwrap_err();
-
-    let diagnostic = error.diagnostic().unwrap();
-    assert_eq!(diagnostic.phase, SqliteFailurePhase::SourceValidation);
-    assert_eq!(diagnostic.artifact, SqliteArtifactKind::ProviderDatabase);
-    assert!(matches!(
-        diagnostic.sqlite_primary_code,
-        Some(ffi::SQLITE_CORRUPT) | Some(ffi::SQLITE_NOTADB)
-    ));
-    assert_eq!(diagnostic.retry, SqliteRetryDecision::DoNotRetryCorrupt);
-    assert_eq!(
-        parent.snapshot_counters().source_bytes_copied(),
-        expected_copy_bytes
-    );
-    assert_eq!(directory_file_bytes(temp.path()), before);
-}
-
-#[test]
-fn malformed_private_source_copy_is_ctx_owned_and_not_retried() {
-    let temp = tempfile::tempdir().unwrap();
-    let database = temp.path().join("provider.sqlite");
-    let _writer = create_persistent_wal(&database);
-    let expected_copy_bytes = fs::metadata(&database).unwrap().len()
-        + fs::metadata(database.with_file_name("provider.sqlite-wal"))
-            .unwrap()
-            .len();
-    let before = directory_file_bytes(temp.path());
-    let parent = retain_parent(temp.path());
-
-    let error = open_root_handle_sqlite_source_online_backup_after_private_source_copy_for_test(
-        &parent,
-        OsStr::new("provider.sqlite"),
-        |source_copy| {
-            let mut copy = OpenOptions::new().write(true).open(source_copy).unwrap();
-            copy.seek(SeekFrom::Start(0)).unwrap();
-            copy.write_all(&[0_u8; 100]).unwrap();
-            copy.sync_all().unwrap();
-        },
-    )
-    .unwrap_err();
-
-    let diagnostic = error.diagnostic().unwrap();
-    assert_eq!(diagnostic.phase, SqliteFailurePhase::SourceValidation);
-    assert_eq!(diagnostic.artifact, SqliteArtifactKind::PrivateSourceCopy);
-    assert!(matches!(
-        diagnostic.sqlite_primary_code,
-        Some(ffi::SQLITE_CORRUPT) | Some(ffi::SQLITE_NOTADB)
-    ));
-    assert_eq!(diagnostic.retry, SqliteRetryDecision::DoNotRetryCorrupt);
-    assert!(error.is_ctx_owned_corruption());
-    assert_eq!(
-        parent.snapshot_counters().source_bytes_copied(),
-        expected_copy_bytes
-    );
-    assert_eq!(directory_file_bytes(temp.path()), before);
-}
-
-#[test]
-fn malformed_private_backup_is_distinguished_from_source_corruption() {
-    let temp = tempfile::tempdir().unwrap();
-    let database = temp.path().join("provider.sqlite");
-    create_database(&database, "expected");
-    let before = directory_file_bytes(temp.path());
-    let parent = retain_parent(temp.path());
-
-    let error = open_root_handle_sqlite_source_online_backup_after_backup_for_test(
-        &parent,
-        OsStr::new("provider.sqlite"),
-        |backup_path| {
-            let mut backup = OpenOptions::new().write(true).open(backup_path).unwrap();
-            backup.seek(SeekFrom::Start(0)).unwrap();
-            backup.write_all(&[0_u8; 100]).unwrap();
-            backup.sync_all().unwrap();
-        },
-    )
-    .unwrap_err();
-
-    let diagnostic = error.diagnostic().unwrap();
-    assert_eq!(diagnostic.phase, SqliteFailurePhase::BackupValidation);
-    assert_eq!(diagnostic.artifact, SqliteArtifactKind::PrivateBackup);
-    assert!(matches!(
-        diagnostic.sqlite_primary_code,
-        Some(ffi::SQLITE_CORRUPT) | Some(ffi::SQLITE_NOTADB)
-    ));
-    assert_eq!(diagnostic.retry, SqliteRetryDecision::DoNotRetryCorrupt);
-    assert!(diagnostic.copied_pages > 0);
-    assert!(diagnostic.copied_bytes > 0);
-    assert_eq!(directory_file_bytes(temp.path()), before);
-}
-
-#[test]
-fn private_source_copy_cleanup_failure_is_explicit_and_route_fatal() {
-    let temp = tempfile::tempdir().unwrap();
-    let database = temp.path().join("provider.sqlite");
-    let _writer = create_persistent_wal(&database);
-    let parent = retain_parent(temp.path());
-    let moved = temp.path().join("retained-source-copy-for-cleanup-test");
-
-    let error = open_root_handle_sqlite_source_online_backup_after_private_source_copy_for_test(
-        &parent,
-        OsStr::new("provider.sqlite"),
-        |source_copy| {
-            let directory = source_copy.parent().unwrap();
-            fs::rename(directory, &moved).unwrap();
-            fs::write(directory, b"blocks directory cleanup").unwrap();
-        },
-    )
-    .unwrap_err();
-
-    let diagnostic = error.diagnostic().unwrap();
-    assert_eq!(diagnostic.phase, SqliteFailurePhase::Cleanup);
-    assert_eq!(diagnostic.artifact, SqliteArtifactKind::PrivateSourceCopy);
-    assert_eq!(diagnostic.cleanup, SqliteCleanupStatus::Failed);
-    assert_eq!(diagnostic.retry, SqliteRetryDecision::RouteFatalResource);
-    assert!(error.is_systemic_resource_failure());
-    fs::remove_file(error_cleanup_path(&error)).unwrap();
-    fs::remove_dir_all(moved).unwrap();
-}
-
-#[test]
-fn private_backup_cleanup_failure_is_explicit_and_route_fatal() {
-    let temp = tempfile::tempdir().unwrap();
-    let database = temp.path().join("provider.sqlite");
-    create_database(&database, "expected");
-    let parent = retain_parent(temp.path());
-    let moved = temp.path().join("retained-backup-for-cleanup-test");
-
-    let error = open_root_handle_sqlite_source_online_backup_after_backup_for_test(
-        &parent,
-        OsStr::new("provider.sqlite"),
-        |backup| {
-            let directory = backup.parent().unwrap();
-            fs::rename(directory, &moved).unwrap();
-            fs::write(directory, b"blocks directory cleanup").unwrap();
-        },
-    )
-    .unwrap_err();
-
-    let diagnostic = error.diagnostic().unwrap();
-    assert_eq!(diagnostic.phase, SqliteFailurePhase::Cleanup);
-    assert_eq!(diagnostic.artifact, SqliteArtifactKind::PrivateBackup);
-    assert_eq!(diagnostic.cleanup, SqliteCleanupStatus::Failed);
-    assert_eq!(diagnostic.retry, SqliteRetryDecision::RouteFatalResource);
-    assert!(error.is_systemic_resource_failure());
-    fs::remove_file(error_cleanup_path(&error)).unwrap();
-    fs::remove_dir_all(moved).unwrap();
-}
-
-fn error_cleanup_path(error: &SqliteSourceAccessError) -> &Path {
-    match error {
-        SqliteSourceAccessError::Diagnosed { source, .. } => error_cleanup_path(source),
-        SqliteSourceAccessError::ScratchIoUnavailable { path, .. } => path,
-        other => panic!("unexpected cleanup error: {other:?}"),
-    }
-}
-
-#[test]
 fn logical_online_backup_progress_is_monotonic_and_terminal() {
     let temp = tempfile::tempdir().unwrap();
     let database = temp.path().join("provider.sqlite");
@@ -1650,96 +1335,6 @@ fn logical_online_backup_rejects_checkpoint_during_copied_family_acquisition() {
     );
 
     assert!(result.unwrap_err().is_source_changed());
-}
-
-#[test]
-fn production_logical_acquisition_retries_commit_inside_family_copy_window() {
-    assert_production_family_copy_race_retries(false);
-}
-
-#[test]
-fn production_logical_acquisition_retries_checkpoint_inside_family_copy_window() {
-    assert_production_family_copy_race_retries(true);
-}
-
-fn assert_production_family_copy_race_retries(checkpoint: bool) {
-    let temp = tempfile::tempdir().unwrap();
-    let database = temp.path().join("provider.sqlite");
-    let writer = create_large_persistent_wal(&database);
-    let shared_memory = database.with_file_name("provider.sqlite-shm");
-    assert!(shared_memory.exists());
-    let parent = retain_parent(temp.path());
-    let mut settled_provider_state = None;
-    let mut attempt_starts = 0_u32;
-
-    let snapshot = parent
-        .open_logical_online_backup_snapshot_with_progress(
-            OsStr::new("provider.sqlite"),
-            |progress| {
-                if progress.stage == SourceBackedCurrentSourceProgressStage::SourceFamilyCopy {
-                    if progress.snapshot_bytes_completed == Some(0) {
-                        attempt_starts += 1;
-                    }
-                    let inside_copy = progress
-                        .snapshot_bytes_completed
-                        .zip(progress.snapshot_bytes_total)
-                        .is_some_and(|(completed, total)| completed > 0 && completed < total);
-                    if inside_copy && settled_provider_state.is_none() {
-                        if checkpoint {
-                            let (busy, _, _): (i64, i64, i64) = writer
-                                .query_row("PRAGMA wal_checkpoint(TRUNCATE)", [], |row| {
-                                    Ok((row.get(0)?, row.get(1)?, row.get(2)?))
-                                })
-                                .unwrap();
-                            assert_eq!(busy, 0, "the checkpoint fixture must complete");
-                            assert_eq!(
-                                fs::metadata(database.with_file_name("provider.sqlite-wal"))
-                                    .unwrap()
-                                    .len(),
-                                0,
-                                "the checkpoint fixture must truncate the WAL"
-                            );
-                        } else {
-                            writer
-                                .execute("INSERT INTO messages (body) VALUES ('raced-commit')", [])
-                                .unwrap();
-                        }
-                        settled_provider_state = Some(directory_file_bytes(temp.path()));
-                    }
-                }
-                Ok::<(), std::convert::Infallible>(())
-            },
-        )
-        .unwrap();
-
-    assert_eq!(
-        parent
-            .snapshot_counters()
-            .logical_source_transition_retries(),
-        1,
-        "the production path must retry exactly once"
-    );
-    assert_eq!(
-        attempt_starts,
-        if checkpoint { 1 } else { 2 },
-        "the second checkpoint attempt uses the immutable main path after WAL truncation"
-    );
-    let settled_provider_state = settled_provider_state.unwrap();
-    for expected in [
-        "provider.sqlite",
-        "provider.sqlite-wal",
-        "provider.sqlite-shm",
-    ] {
-        assert!(settled_provider_state.contains_key(OsStr::new(expected)));
-    }
-    let values = read_values(&snapshot);
-    if checkpoint {
-        assert_eq!(values, ["before-wal", "from-wal"]);
-    } else {
-        assert_eq!(values, ["before-wal", "from-wal", "raced-commit"]);
-    }
-    snapshot.finish().unwrap();
-    assert_eq!(directory_file_bytes(temp.path()), settled_provider_state);
 }
 
 #[cfg(unix)]
